requirements.txt
fastapi
uvicorn
python-multipart
google-genai
faiss-cpu
sentence-transformers
numpy
tqdm
pdfplumber
python-docx
python-pptx
pydantic
